xml2Parquet/requirement.txt
pandas	2.0.3
pillow	10.4.0
pyarrow	17.0.0
sklearn	0.0
